Fix wrong slicing of a list of 1 dim ndarrays
`img = img_line[:][0:cols]` was not slicing to the number of
columns but was slicing the rows in stead. It's now re-written
to use a 2-dim numpy array iso a python list of 1-dim numpy arrays.
This 2-dim numpy array uses the same base array the `img` to save
memory and it is sliced to the correct size before being passed to
code that does the actual FFT calculation.