Fix wrong slicing of a list of 1 dim ndarrays (#148)
* Fix wrong slicing of a list of 1 dim ndarrays

`img = img_line[:][0:cols]` was not slicing to the number of
columns but was slicing the rows in stead. It's now re-written
to use a 2-dim numpy array iso a python list of 1-dim numpy arrays.
This 2-dim numpy array uses the same base array the `img` to save
memory and it is sliced to the correct size before being passed to
code that does the actual FFT calculation.

* Add test for masked fft calculation

It fails on the previous implementation which had some wrong slicing
in it.

* Use regular copy i.s.o. masked copy

Otherwise the result will always be a masked_array instance and the
`masked route` will be taken even when it's not needed.

* Outsource row wise compress of a masked array to a separate function

Also renamed argument img_ to img. Not sure what that was for.

Co-authored-by: Louver <lvervecken@industrialise.be>